+++ apps/search/management/commands/esreindex.py
@@ -1,36 +1,9 @@
-import time
 from optparse import make_option
 
 from django.core.management.base import BaseCommand, CommandError
 
 from search.es_utils import es_reindex_cmd
-
-
-class FakeLogger(object):
-    """Fake logger that we can pretend is a Python Logger
-
-    Why? Well, because Django has logging settings that prevent me
-    from setting up a logger here that uses the stdout that the Django
-    BaseCommand has. At some point p while fiddling with it, I
-    figured, 'screw it--I'll just write my own' and did.
-
-    The minor ramification is that this isn't a complete
-    implementation so if it's missing stuff, we'll have to add it.
-    """
-
-    def __init__(self, stdout):
-        self.stdout = stdout
-
-    def _out(self, level, msg, *args):
-        msg = msg % args
-        self.stdout.write('%s %-8s: %s\n' % (
-                time.strftime('%H:%M:%S'), level, msg))
-
-    def info(self, msg, *args):
-        self._out('INFO', msg, *args)
-
-    def error(self, msg, *args):
-        self._out('ERROR', msg, *args)
+from search.utils import FakeLogger
 
 
 class Command(BaseCommand):
